requirements.txt
pynetbox
pyyaml
